nhl_projection_starter/requirements.txt
pandas>=2.2.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
python-dotenv>=1.0.1
pyarrow>=18.0.0
jupyter>=1.1.1
ipykernel>=6.29.5
google-api-python-client>=2.169.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
gspread>=6.1.4
duckdb>=1.1.3
